tidied

+++ Assignment3/Test_lucas.m
@@ -1,3 +1,3 @@
-img = rgb2gray(imread('sphere1.ppm'));
+img = double(rgb2gray(imread('sphere1.ppm')));
 size(img);
 I = lucas_kanade(img);
